requirements.txt
streamlit
pandas
google-generativeai
wordcloud
matplotlib
openpyxl
Emoji